backend/requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2

# LLM
google-genai==1.5.0

# LiveKit
livekit-agents==1.0.0
livekit-plugins-google==1.0.0
livekit-plugins-silero==1.0.0
livekit==0.17.5

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9

# Celery + Redis
celery==5.4.0
redis==5.1.1

# Utils
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
